refactor(redos): extract the structural primitives and rules
guard_core/detection_engine/_redos_structure.py was rank B (MI 13.78).
The char-class/group/alternation walking primitives (_skip_char_class,
_find_group_end, _iter_quantified_group_bodies, and friends) move to
_redos_structure_primitives.py; the two detection entry points
(_detect_nested_unbounded_quantifier, _detect_adjacent_broad_unbounded_quantifiers)
and their broad-shorthand helpers move to _redos_structure_rules.py.
_redos_structure.py re-exports every name it exported before; nothing
importable moved. Verified byte-identical against the pre-split module
across a range of test patterns.

Sync mirrors regenerated.
